+++ src/main/java/roomescape/dao/TimeDAO.java
@@ -8,7 +8,7 @@ import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.model.domain.Time;
+import roomescape.domain.Time;
 import roomescape.exception.NotFoundException;
 
 @Repository
